+++ Frequentest-Analysis-CMS-master/CMS-Fequentest/Hyper_Plane.cpp
@@ -3,7 +3,7 @@
 #include "CMS-Frequentest-Analysis.h"
 
 
-Hyper_Plane::Hyper_Plane(std::vector<std::vector<float>>& adataSet)
+Hyper_Plane::Hyper_Plane(std::vector<std::vector<double>>& adataSet)
 {
 	basePlane = true;
 	const int SIZE_OF_DATA_SET = adataSet.size();
@@ -13,7 +13,7 @@ Hyper_Plane::Hyper_Plane(std::vector<std::vector<float>>& adataSet)
 	{
 		points.push_back(find_Smallest_By_Index(adataSet, i));
 	}
-	std::vector<float> orthogonalVector = gaussian_Elimination(points);
+	std::vector<double> orthogonalVector = gaussian_Elimination(points);
 	bool org;
 	for (std::size_t i = 0; i < SIZE_OF_DATA_SET; ++i)
 	{
@@ -37,15 +37,15 @@ Hyper_Plane::Hyper_Plane(std::vector<std::vector<float>>& adataSet)
 	}
 }
 
-Hyper_Plane::Hyper_Plane(std::vector<std::vector<float>>& adataSet,
-	std::vector<std::vector<float>> base_Points, Hyper_Plane base)
+Hyper_Plane::Hyper_Plane(std::vector<std::vector<double>>& adataSet,
+	std::vector<std::vector<double>> base_Points, Hyper_Plane base)
 {
 	basePlane = false;
 	dataSet = adataSet;
 	points = base_Points;
 	const int SIZE_OF_DATA_SET = adataSet.size();
 	const int NUMBER_OF_DIMENSIONS = adataSet[0].size();
-	std::vector<float> orthogonalVector = gaussian_Elimination(points);
+	std::vector<double> orthogonalVector = gaussian_Elimination(points);
 	if (find_Smallest_By_Index(base_Points, 0)[0] < base.smallest_Point()[0])
 	{
 		if (orthogonalVector[0] > 0)
@@ -88,16 +88,16 @@ Hyper_Plane::Hyper_Plane(std::vector<std::vector<float>>& adataSet,
 	}
 }
 
-std::vector<float> Hyper_Plane::smallest_Point()
+std::vector<double> Hyper_Plane::smallest_Point()
 {
 	return find_Smallest_By_Index(points, 0);
 }
 
 
-bool Hyper_Plane::check_Point_Outside(std::vector<float> point)
+bool Hyper_Plane::check_Point_Outside(std::vector<double> point)
 {
 	const int NUMBER_OF_DIMENSIONS = dataSet[0].size();
-	float sum = 0;
+	double sum = 0;
 	for (std::size_t i = 0; i < NUMBER_OF_DIMENSIONS; ++i)
 	{
 		sum += point[i] * orthogonalVector[i];
@@ -126,7 +126,7 @@ bool Hyper_Plane::check_Point_Outside(std::vector<float> point)
 	}
 }
 
-double Hyper_Plane::dist_To_Point(std::vector<float> point)
+double Hyper_Plane::dist_To_Point(std::vector<double> point)
 {
 	const int NUMBER_OF_DIMENSIONS = orthogonalVector.size();
 	double numerator = 0;
@@ -149,8 +149,8 @@ double Hyper_Plane::dist_To_Point(std::vector<float> point)
 std::vector<Hyper_Plane> Hyper_Plane::expand_Surface()
 {
 	const int SIZE_OF_DATA_SET = dataSet.size();
-	float dist = 0;
-	std::vector<float> point;
+	double dist = 0;
+	std::vector<double> point;
 	std::vector<Hyper_Plane> out;
 	for (std::size_t i = 0; i < SIZE_OF_DATA_SET; ++i)
 	{
@@ -163,7 +163,7 @@ std::vector<Hyper_Plane> Hyper_Plane::expand_Surface()
 			}
 		}
 	}
-	std::vector<std::vector<float>> outPoints;
+	std::vector<std::vector<double>> outPoints;
 	for (std::size_t i = 0; i < points.size(); ++i)
 	{
 		for (std::size_t j = 0; j < points.size(); ++j)
@@ -175,7 +175,7 @@ std::vector<Hyper_Plane> Hyper_Plane::expand_Surface()
 		}
 		outPoints.push_back(point);
 		out.push_back(Hyper_Plane(dataSet, outPoints, *this));
-		std::vector<std::vector<float>> outPoints;
+		std::vector<std::vector<double>> outPoints;
 	}
 	delete this;
 	return out;
