+++ Frequentest-Analysis-CMS-master/CMS-Fequentest/Hyper_Surface.h
@@ -4,19 +4,19 @@
 class Hyper_Surface
 {
 private:
-	std::vector<std::vector<float>> dataSet;
+	std::vector<std::vector<double>> dataSet;
 	std::vector<Hyper_Plane> planeSet;
 	const int NUMBER_OF_DATA_POINTS = dataSet.size();
 	const int NUMBER_OF_PLANES = planeSet.size();
 public:
-	Hyper_Surface(std::vector<std::vector<float>> adataSet);
+	Hyper_Surface(std::vector<std::vector<double>> adataSet);
 	~Hyper_Surface();
 	// Checks if there are more points outside of the surface
 	bool check_If_More();
 	// Makes the whole Hyper_Surface
 	void make_Surface();
 	// Checks if a point is in the surface
-	bool point_Is_In(std::vector<float> point);
+	bool point_Is_In(std::vector<double> point);
 	// Cleares all the points inside of the Hyper_Surface
 	void clear_Points();
 };
